requirements.txt
flask
google-generativeai
python-dotenv
requests